requirements.txt
astroid==1.4.5
colorama==0.3.7
future==0.15.2
git-pylint-commit-hook==2.1.1
greenlet==0.4.9
lazy-object-proxy==1.2.1
pylint==1.5.5
python-memcached==1.57
requests==2.9.1
six==1.10.0
wheel==0.24.0
wrapt==1.10.6
